requirements.txt
google-cloud-firestore==2.7.2
parameterized==0.8.1
mock-firestore==0.11.0
functions_framework==3.2.1